validation helper functions called properly
