+++ src/lib/validation-functions.ts
@@ -207,3 +207,49 @@ export function ensureAssistantsAreAssignedToOnlyOneClass(
     },
   };
 }
+
+export function ensureAllAvailableAssistantsAreAssigned(
+  assignments: Allocation[],
+  plaAvailableSet: Set<string>,
+  taAvailableSet: Set<string>,
+): ValidationResult {
+  const t0 = performance.now();
+  const errors: string[] = [];
+  const warnings: string[] = [];
+
+  const assignedPLAs = new Set<string>();
+  const assignedTAs = new Set<string>();
+
+  for (const assignment of assignments) {
+    for (const pla of assignment.PLAs) {
+      assignedPLAs.add(personKey(pla));
+    }
+    for (const ta of assignment.TAs) {
+      assignedTAs.add(personKey(ta));
+    }
+  }
+
+  for (const pla of plaAvailableSet) {
+    if (!assignedPLAs.has(pla)) {
+      warnings.push(
+        `PLA "${pla}" is available but not assigned to any course.`,
+      );
+    }
+  }
+
+  for (const ta of taAvailableSet) {
+    if (!assignedTAs.has(ta)) {
+      errors.push(`TA "${ta}" is available but not assigned to any course.`);
+    }
+  }
+
+  return {
+    ok: errors.length === 0,
+    errors,
+    warnings,
+    meta: {
+      ms: Math.round(performance.now() - t0),
+      rule: "All available assistants should be assigned at least once.",
+    },
+  };
+}
